Rename seedu.duke to seedu.navi
