+++ docs/AboutUs.md
@@ -6,5 +6,3 @@
 | ![](https://via.placeholder.com/100.png?text=Photo) |     Lakshmi     |           [lakshm1i](https://github.com/lakshm1i)           |      [Portfolio](docs/team/lakshm1i.md)      |
 | ![](https://via.placeholder.com/100.png?text=Photo) |    Keerthana    | [KeerthanaBanukumar](https://github.com/KeerthanaBanukumar) | [Portfolio](docs/team/KeerthanaBanukumar.md) |
 | ![](https://via.placeholder.com/100.png?text=Photo) | Chuan-Kai Chang |      [kevinchangckc](https://github.com/kevinchangckc)      |   [Portfolio](docs/team/kevinchangckc.md)    |
-
-
