+++ src/main/java/seedu/navi/Navi.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.navi;
 
 import java.util.Scanner;
